fix en printPoint, ya que al ser Input tipo double y la funcion lo tomaba como un uint16_t, no se graficaba correctamente

+++ src/main.cpp
@@ -186,7 +186,7 @@ void loop() {
             }
 
             // Voy gráficando la evolución de la temperatura con el tiempo
-            printPoint(tiempo, Input1);
+            printPoint(tiempo, (uint16_t)Input1);
         }
         else{
             perfilRamp = 0;
